move custom fields out of payloads.
