+++ hardware/opentrons_hardware/firmware_bindings/messages/payloads.py
@@ -3,10 +3,11 @@
 #  dataclass fields interpretation.
 #  from __future__ import annotations
 from dataclasses import dataclass
-import enum
 
-from .. import utils, ErrorCode
-from ..constants import SensorType, ToolType
+from .fields import FirmwareShortSHADataField, VersionFlagsField, \
+    TaskNameDataField, ToolField, FirmwareUpdateDataField, ErrorCodeField, \
+    SensorTypeField
+from .. import utils
 
 
 @dataclass
@@ -16,38 +17,6 @@ class EmptyPayload(utils.BinarySerializable):
     pass
 
 
-class FirmwareShortSHADataField(utils.BinaryFieldBase[bytes]):
-    """The short hash in a device info.
-
-    This is sized to hold the default size of an abbreviated Git hash,
-    what you get when you do git rev-parse --short HEAD. If we ever
-    need to increase the size of that abbreviated ID, we'll need to
-    increase this too.
-    """
-
-    NUM_BYTES = 7
-    FORMAT = f"{NUM_BYTES}s"
-
-
-class VersionFlags(enum.Enum):
-    """Flags in the version field."""
-
-    BUILD_IS_EXACT_COMMIT = 0x1
-    BUILD_IS_EXACT_VERSION = 0x2
-    BUILD_IS_FROM_CI = 0x4
-
-
-class VersionFlagsField(utils.UInt32Field):
-    """A field for version flags."""
-
-    def __repr__(self) -> str:
-        """Print version flags."""
-        flags_list = [
-            flag.name for flag in VersionFlags if bool(self.value & flag.value)
-        ]
-        return f"{self.__class__.__name__}(value={','.join(flags_list)})"
-
-
 @dataclass
 class DeviceInfoResponsePayload(utils.BinarySerializable):
     """Device info response."""
@@ -55,13 +24,6 @@ class DeviceInfoResponsePayload(utils.BinarySerializable):
     version: utils.UInt32Field
     flags: VersionFlagsField
     shortsha: FirmwareShortSHADataField
-
-
-class TaskNameDataField(utils.BinaryFieldBase[bytes]):
-    """The name field of TaskInfoResponsePayload."""
-
-    NUM_BYTES = 12
-    FORMAT = f"{NUM_BYTES}s"
 
 
 @dataclass
@@ -225,18 +187,6 @@ class ReadPresenceSensingVoltageResponsePayload(utils.BinarySerializable):
     gripper: utils.UInt16Field
 
 
-class ToolField(utils.UInt8Field):
-    """A tool field."""
-
-    def __repr__(self) -> str:
-        """Print out a tool string."""
-        try:
-            tool_val = ToolType(self.value).name
-        except ValueError:
-            tool_val = str(self.value)
-        return f"{self.__class__.__name__}(value={tool_val})"
-
-
 @dataclass
 class ToolsDetectedNotificationPayload(utils.BinarySerializable):
     """Tool detection notification."""
@@ -252,13 +202,6 @@ class FirmwareUpdateWithAddress(utils.BinarySerializable):
     """A FW update payload with an address."""
 
     address: utils.UInt32Field
-
-
-class FirmwareUpdateDataField(utils.BinaryFieldBase[bytes]):
-    """The data field of FirmwareUpdateData."""
-
-    NUM_BYTES = 56
-    FORMAT = f"{NUM_BYTES}s"
 
 
 @dataclass
@@ -295,18 +238,6 @@ class FirmwareUpdateData(FirmwareUpdateWithAddress):
         return obj
 
 
-class ErrorCodeField(utils.UInt16Field):
-    """Error code field."""
-
-    def __repr__(self) -> str:
-        """Print error code."""
-        try:
-            error = ErrorCode(self.value).name
-        except ValueError as e:
-            error = str(self.value)
-        return f"{self.__class__.__name__}(value={error})"
-
-
 @dataclass
 class FirmwareUpdateDataAcknowledge(FirmwareUpdateWithAddress):
     """A FW update data acknowledge payload."""
@@ -341,18 +272,6 @@ class GetLimitSwitchResponse(utils.BinarySerializable):
     """A response to the Limit Switch Status request payload."""
 
     switch_status: utils.UInt8Field
-
-
-class SensorTypeField(utils.UInt8Field):
-    """sensor type."""
-
-    def __repr__(self) -> str:
-        """Print sensor."""
-        try:
-            sensor_val = SensorType(self.value).name
-        except ValueError:
-            sensor_val = str(self.value)
-        return f"{self.__class__.__name__}(value={sensor_val})"
 
 
 @dataclass
