+++ hardware/tests/opentrons_hardware/firmware_bindings/test_messages/test_payloads.py
@@ -1,6 +1,7 @@
 """Payloads tests."""
 import pytest
 
+import opentrons_hardware.firmware_bindings.messages.fields
 from opentrons_hardware.firmware_bindings.messages import payloads
 from opentrons_hardware.firmware_bindings import utils
 
@@ -23,6 +24,6 @@ def test_create_firmware_updata_data(
         address=utils.UInt32Field(address),
         num_bytes=utils.UInt8Field(len(data)),
         reserved=utils.UInt8Field(0),
-        data=payloads.FirmwareUpdateDataField(data),
+        data=opentrons_hardware.firmware_bindings.messages.fields.FirmwareUpdateDataField(data),
         checksum=utils.UInt16Field(expected_checksum),
     )
