+++ hardware/tests/opentrons_hardware/firmware_update/test_eraser.py
@@ -13,7 +13,7 @@ from opentrons_hardware.firmware_bindings.messages import (
     message_definitions,
     payloads,
 )
-from opentrons_hardware.firmware_bindings.messages.payloads import ErrorCodeField
+from opentrons_hardware.firmware_bindings.messages.fields import ErrorCodeField
 from opentrons_hardware.firmware_bindings.utils import UInt16Field
 from opentrons_hardware.firmware_update import FirmwareUpdateEraser
 from opentrons_hardware.firmware_update.errors import ErrorResponse, TimeoutResponse
