+++ hardware/tests/opentrons_hardware/firmware_update/test_initiater.py
@@ -1,6 +1,8 @@
 """Tests for FirmwareUpdateInitiator."""
 import pytest
 from mock import AsyncMock, call
+
+import opentrons_hardware.firmware_bindings.messages.fields
 from opentrons_hardware.firmware_bindings import (
     NodeId,
     ArbitrationId,
@@ -34,8 +36,8 @@ async def test_messaging(
             response = message_definitions.DeviceInfoResponse(
                 payload=payloads.DeviceInfoResponsePayload(
                     version=UInt32Field(0),
-                    flags=payloads.VersionFlagsField(0),
-                    shortsha=payloads.FirmwareShortSHADataField(b"abcdef0"),
+                    flags=opentrons_hardware.firmware_bindings.messages.fields.VersionFlagsField(0),
+                    shortsha=opentrons_hardware.firmware_bindings.messages.fields.FirmwareShortSHADataField(b"abcdef0"),
                 )
             )
             can_message_notifier.notify(
@@ -82,8 +84,8 @@ async def test_retry(
         message_definitions.DeviceInfoResponse(
             payload=payloads.DeviceInfoResponsePayload(
                 version=UInt32Field(0),
-                flags=payloads.VersionFlagsField(0),
-                shortsha=payloads.FirmwareShortSHADataField(b"abcdef0"),
+                flags=opentrons_hardware.firmware_bindings.messages.fields.VersionFlagsField(0),
+                shortsha=opentrons_hardware.firmware_bindings.messages.fields.FirmwareShortSHADataField(b"abcdef0"),
             )
         ),
         None,
