+++ hardware/tests/opentrons_hardware/hardware_control/tools/test_detector.py
@@ -6,7 +6,7 @@ from opentrons_hardware.firmware_bindings.messages import (
 import pytest
 from mock import AsyncMock, call
 
-from opentrons_hardware.firmware_bindings.messages.payloads import ToolField
+from opentrons_hardware.firmware_bindings.messages.fields import ToolField
 from opentrons_hardware.hardware_control.tools import detector
 from opentrons_hardware.firmware_bindings.utils import UInt8Field
 from opentrons_hardware.firmware_bindings.messages import message_definitions, payloads
